NETPYNE-209 hooked events on notebook load
